# openapi.yaml
openapi: 3.1.0
info:
  title: Email Validator API
  version: 1.0.0
  description: API for validating email addresses and providing suggestions for typos

servers:
  - url: /api

paths:
  /validate:
    get:
      summary: Validate a single email address
      description: Validates an email address and returns detailed information about its validity
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: Successful validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResult'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Validate a single email address
      description: Validates an email address and returns detailed information about its validity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmailValidationRequest'
      responses:
        '200':
          description: Successful validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResult'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /validate/batch:
    get:
      summary: Validate multiple email addresses
      description: Validates multiple email addresses in a single request
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: array
            items:
              type: string
              format: email
          style: form
          explode: true
      responses:
        '200':
          description: Successful validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchValidationResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Validate multiple email addresses
      description: Validates multiple email addresses in a single request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchValidationRequest'
      responses:
        '200':
          description: Successful validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchValidationResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /typo-suggestions:
    get:
      summary: Get typo suggestions for an email address
      description: Returns suggestions for possible typos in the email address
      parameters:
        - name: email
          in: query
          required: true
          schema:
            type: string
            format: email
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TypoSuggestionResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Get typo suggestions for an email address
      description: Returns suggestions for possible typos in the email address
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TypoSuggestionRequest'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TypoSuggestionResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /status:
    get:
      summary: Get API status
      description: Returns the current status of the API
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIStatus'

components:
  schemas:
    ValidationResult:
      type: object
      properties:
        email:
          type: string
          format: email
          description: The email address that was validated
        validations:
          type: object
          properties:
            syntax:
              type: boolean
              description: Whether the email has valid syntax
            domain_exists:
              type: boolean
              description: Whether the domain exists
            mx_records:
              type: boolean
              description: Whether the domain has valid MX records
            mailbox_exists:
              type: boolean
              description: Whether the mailbox exists
            is_disposable:
              type: boolean
              description: Whether the email is from a disposable provider
            is_role_based:
              type: boolean
              description: Whether the email is a role-based address
        score:
          type: integer
          minimum: 0
          maximum: 100
          description: Overall validation score
        status:
          type: string
          enum:
            - VALID
            - PROBABLY_VALID
            - INVALID
            - MISSING_EMAIL
            - INVALID_FORMAT
            - INVALID_DOMAIN
            - NO_MX_RECORDS
            - DISPOSABLE
          description: Validation status

    EmailValidationRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          description: The email address to validate

    BatchValidationRequest:
      type: object
      required:
        - emails
      properties:
        emails:
          type: array
          items:
            type: string
            format: email
          description: List of email addresses to validate

    BatchValidationResponse:
      type: object
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/ValidationResult'
          description: List of validation results

    TypoSuggestionRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          description: The email address to check for typos

    TypoSuggestionResponse:
      type: object
      properties:
        email:
          type: string
          format: email
          description: The original email address
        suggestions:
          type: array
          items:
            type: string
          description: List of suggested corrections

    APIStatus:
      type: object
      properties:
        status:
          type: string
          description: Current status of the API
        uptime:
          type: string
          description: Time since the API started
        requests_handled:
          type: integer
          description: Total number of requests handled
        average_response_time_ms:
          type: number
          description: Average response time in milliseconds

    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message describing what went wrong 